split Network class implementation
